feat: add ChatController and HomeController for NL2SQL functionality
