--- spring-ai-alibaba-nl2sql-example/chat/src/main/java/com.alibaba.cloud.ai.example/HomeController.java
@@ -1,17 +1,0 @@
-package com.alibaba.cloud.ai.example;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-
-@Controller
-public class HomeController {
-
-    @GetMapping("/")
-    @PostMapping("/")
-    public String index() {
-        return "index";
-    }
-
-
-}
