+++ spring-ai-alibaba-nl2sql-example/chat/src/main/resources/static/index.html
@@ -67,15 +67,15 @@
 
         .input-group textarea {
             width: 100%;
-            padding: 20px;
+            padding: 24px;
             border: 2px solid #e1e8ed;
-            border-radius: 15px;
-            font-size: 16px;
+            border-radius: 18px;
+            font-size: 18px;
             resize: vertical;
-            min-height: 140px;
+            min-height: 200px;
             transition: all 0.3s ease;
             font-family: inherit;
-            line-height: 1.5;
+            line-height: 1.7;
         }
 
         .input-group textarea:focus {
@@ -265,14 +265,19 @@
         <div class="query-section">
             <h2>📝 输入您的查询</h2>
             <div class="input-group">
-                    <textarea
-                            id="queryInput"
-                            placeholder="请输入您想查询的问题，例如：
-• 销售额最高的产品是什么？
-• 上个月的用户增长情况如何？
-• 哪个地区的订单量最多？
-• 客户满意度趋势分析..."
-                    ></textarea>
+                <label style="font-weight:600;margin-bottom:8px;display:block;">选择向量库类型：</label>
+                <select id="vectorTypeSelect" style="margin-bottom:15px;padding:8px 12px;border-radius:8px;border:1px solid #e1e8ed;font-size:15px;">
+                    <option value="memory">Memory(SimpleVector)</option>
+                    <option value="analyticDb">AnalyticDB</option>
+                </select>
+                <textarea
+                        id="queryInput"
+                        placeholder="请输入您想查询的问题，例如：
+                        • 销售额最高的产品是什么？
+                        • 上个月的用户增长情况如何？
+                        • 哪个地区的订单量最多？
+                        • 客户满意度趋势分析..."
+                ></textarea>
             </div>
             <button id="queryBtn" class="btn-primary">
                 🚀 开始查询
@@ -299,6 +304,7 @@
 
     // 查询函数
     async function performQuery(input) {
+        const vectorType = document.getElementById('vectorTypeSelect').value;
         if (!input.trim()) {
             alert('请输入查询内容');
             return;
@@ -308,7 +314,8 @@
         setLoading(true);
 
         try {
-            const response = await fetch('http://localhost:8065/chat', {
+            const apiUrl = vectorType === 'analyticDb' ? 'http://localhost:8066/chat' : 'http://localhost:8066/simpleChat';
+            const response = await fetch(apiUrl, {
                 method: 'POST',
                 headers: {
                     'Content-Type': 'application/json',
