+++ spring-ai-alibaba-nl2sql-example/mcp/src/main/java/com/alibaba/cloud/ai/example/service/McpService.java
@@ -15,7 +15,8 @@
  */
 package com.alibaba.cloud.ai.example.service;
 
-import com.alibaba.cloud.ai.service.Nl2SqlService;
+import com.alibaba.cloud.ai.service.analytic.AnalyticNl2SqlService;
+import com.alibaba.cloud.ai.service.simple.SimpleNl2SqlService;
 import org.springframework.ai.tool.annotation.Tool;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -24,7 +25,10 @@ import org.springframework.stereotype.Service;
 public class McpService {
 
 	@Autowired
-	private Nl2SqlService nl2SqlService;
+	private AnalyticNl2SqlService nl2SqlService;
+
+	@Autowired
+	private SimpleNl2SqlService simpleNl2SqlService;
 
 	/**
 	 * 从数据库中获取问题所需要的数据
@@ -34,6 +38,12 @@ public class McpService {
 	public String nl2Sql(String input) throws Exception {
 		String sql = nl2SqlService.nl2sql(input);
 		return nl2SqlService.executeSql(sql);
+	}
+
+	@Tool(description = "使用内存向量库从数据库中获取问题所需要数据")
+	public String simpleNl2Sql(String input) throws Exception {
+		String sql = simpleNl2SqlService.nl2sql(input);
+		return simpleNl2SqlService.executeSql(sql);
 	}
 
 }
